tracker/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "snoke-tracker"
version = "0.1.0"
description = "Local desktop activity tracker for the Snoke weaning app"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pynput>=1.7.7",
    "python-xlib>=0.33; platform_system == 'Linux'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "ruff>=0.7",
]

[project.scripts]
snoke-tracker = "snoke_tracker.__main__:main"

[tool.setuptools.packages.find]
include = ["snoke_tracker*"]
